refactor: RacingGame 내에 입출력 책임 분리
